Temporary Override for Shipyard memcache
The memcache P.S. for Shipyard has been merged a few days ago.
As the OpenStack P.S. to fix the bug in KeystoneMiddleWare is
still under review at the moment, we will need to disable the
'memcache_security_strategy' in Shipyard for now until the bug
is resolved in upstream KeyStone MiddleWare.

Change-Id: I31a9294c06222a8e8fdd8c4aedcb0f197127f97f