Cleanup Airship verbiage in docs
This change attempts to move the airship-in-a-bottle documentation
closer to a target state for Airship by removing references to the
Undercloud Platform. This triggered the reformating of some paragraphs,
but this change does not attempt to correct all of the information -
much more a syntactical change of the documentation than a semantic
change. Requests to correct information in any significant way will be
deferred to subsequent work.

Some completely outdated pages were removed completely, as they provided
more distraction than benefit.

Documents that were "under development" since their inception (and
nothing more than "under development") have been eliminated and may be
re-instantiated if anyone ever has anything to say about the topics, but
for now, removed the clutter.

Change-Id: Ida72b3706c894771888d8b694e1cf95d6cf810ed