Add apparmor profiles on genesis node in multinode gate
The bootactions which will be deployed via Drydock on nodes need
to be performed on Genesis node as well. This should be done as
part of pre-genesis setup before genesis.sh is executed.

This patchset allows apparmor profiles to be staged on the genesis
node in the same manner as seccomp profiles.

Change-Id: I418c955a131dc8e23ab5f4d900ae5ea4f7985468
Reference: https://review.opendev.org/644824