Extended output for pipeline and disk on VM
After timeout pipeline is trying to run deploy_site again and failed.
There is no info in debug report and console output about current status
if the shipyard action. Added new cmd to describe the action.

ceph-mon is detecting low disk space on n0, it's <70%. So, small increase
of disk space is needed.

Change-Id: Ia92390ccadec4485c37038187034a2978911accd